Cart separated as a component
